% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_comparisons_bg2.R
\name{aov2_bg}
\alias{aov2_bg}
\title{Run full 2-way between-groups ANOVA}
\usage{
aov2_bg(
  df,
  DV,
  IV1,
  IV2,
  p.adj = "bonferroni",
  spss.lab = TRUE,
  labels. = NULL
)
}
\arguments{
\item{df}{a data frame with data for the analysis}

\item{IV1, IV2}{the independent variables}

\item{p.adj}{the correction applied to the post-hoc comparisons}

\item{the}{dependent variable}
}
\value{
a list composed of the following objects:
\itemize{
\item Descriptive statistics in an APA-table format
\item Descriptive statistics to plot a graph
\item Table with main and interaction effects results
\item Two tables with post-hoc analyses for the main effects
\item Two tables with simple effects analyses
\item Two tables with post-hoc analyses for the simple effects
}
}
\description{
aov2_bg() runs the complete factorial Analysis of Variance with two independent
variables in a between-objects design. It covers the main and interaction effects,
simple effects, and the post-hoc tests. It also includes descriptives statistics
in an APA-like formatted table. Works in tandem with flex...
}
