% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_modifiers.R
\name{apa_regression}
\alias{apa_regression}
\title{Create an APA formatted regression result}
\usage{
apa_regression(df1, df2, f, p, comma = T)
}
\description{
Create an APA formatted regression result
}
