% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labellers.R
\name{assign_labs}
\alias{assign_labs}
\title{Assign variable labels to a data frame}
\usage{
assign_labs(data, labs)
}
\arguments{
\item{data}{A data frame where the variable labels will be assigned}

\item{labs}{A named character vector where the names correspond to column names in the data frame and the values are the variable labels to be assigned}
}
\value{
The data frame with variable labels assigned
}
\description{
This function assigns variable labels to the columns of a data frame.
The labels are stored as the "label" attribute of the variables.
}
\examples{
data(mtcars)
labs <- list(mpg = "Miles per gallon", wt = "Weight")
assigned_data <- assign_labs(mtcars, labs)

}
