% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_test.R
\name{chi_many}
\alias{chi_many}
\title{Iterate many independent chi-squares over a common variable}
\usage{
chi_many(
  df,
  ...,
  x,
  yes_row = c(1, 2),
  correct = TRUE,
  binary_x = TRUE,
  spss.lab = TRUE,
  labels. = NULL
)
}
\arguments{
\item{df}{a data frame}

\item{...}{binary variables}

\item{x}{second varialbe over which the analysis will be iterated}

\item{yes_row}{which value should be used as the reference for frequencies
and percentages}

\item{correct}{whether to apply continuity correction}

\item{binary_x}{is the IV binary. Otherwise DV is treated as binary.}

\item{spss.lab}{extracts spss-like labels}

\item{labels.}{adds user defined labels}
}
\value{
a data frame with:
\itemize{
\item name of the binary variable
\item frequncies and percentages
\item chi-test results together with Cramer's V results
}
}
\description{
\code{chi_many()} runs many chi-square tests of independence for many binary
y-variables, but always over the same x-variable. Each row contains frequncies
and row percents for one of the two y-variable values, which is selected by
the user.
}
\details{
UWAGA - procenty przy większej liczbie poziomów niż 2 nie mają sensu, o ile
porównywane grupy nie są równoliczne.
}
