% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_test.R
\name{chi_table}
\alias{chi_table}
\title{Compute a cross table with chi-square test}
\usage{
chi_table(
  df,
  x,
  y,
  correct = TRUE,
  spss.lab = TRUE,
  labels. = NULL,
  spss.val = TRUE,
  percents = c("rows", "cols")
)
}
\arguments{
\item{df}{a data frame}

\item{x, y}{atomic vectors (categorical variables)}

\item{correct}{should a continuity correction be applied to chi-square and Cramer's V}

\item{spss.lab}{imports variable labels from SPSS-labelled objects}

\item{labels.}{optional vector containing the variables' labels}

\item{spss.val}{imports value labels from SPSS-labelled objects}

\item{percents}{character value defining the calculation of percents. The default is to "rows".}
}
\value{
a list with two elements
\itemize{
\item character vector with variables' labels to be used in flex_chicross
\item a data frame with cross table like structure
}
}
\description{
\code{chi_table()} creates a neat cross table for two categorical variables with frequencies,
percents, result of the chi-square test of independence, and Cramer's V effect size.
Works in tandem with flex_chicross function from the Cabflex2 package for creating
word-formatted tables.
}
