% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_analysis.R
\name{clean_mediation_results}
\alias{clean_mediation_results}
\title{Cleans the results of mediate_own()}
\usage{
clean_mediation_results(df)
}
\arguments{
\item{df}{A data frame containing the mediation results}
}
\value{
A cleaned data frame with updated columns
}
\description{
This function cleans up the results of a mediation analysis by removing the '(Intercept)' term, rounding the estimates and standard errors to 3 digits, and formatting the results for easy interpretation.
}
\examples{
\dontrun{
df <- mediate_own(data)
cleaned_df <- clean_mediation_results(df)
}
}
\seealso{
\code{\link[=mediate_own]{mediate_own()}}
}
