% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_var}
\alias{compute_var}
\title{Calculate a summarized, index variable}
\usage{
compute_var(
  ...,
  type = c("mean", "sum", "count"),
  count_element = 1,
  na.rm = T
)
}
\arguments{
\item{...}{Atomic vectors}

\item{type}{Type of computation. The default is "mean".}

\item{na.rm}{A logical value (the default is TRUE) indicating whether NA values should be stripped before the
computation proceeds.}

\item{count}{A value to be counted. Used only, if the type is set to "count". The default is set to 1.}
}
\value{
A numeric vector.
}
\description{
\code{compute_var()} creates a variable that is a summary of two or more existing atomic vectors. The calculation can be either
of the three: mean, sum, or count.
}
\details{
When there is only one correct answer that represents a single scorepoint, \code{type = count} is always preferable, as it always
excludes NAs.
}
