% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions_stats.R
\name{conditions_stats}
\alias{conditions_stats}
\title{Compute a pair of grouped APA formatted statistics}
\usage{
conditions_stats(
  df,
  ...,
  IV,
  spss.lab = T,
  labels. = NULL,
  type = c("mean_sd", "median_iqr", "mrank_median")
)
}
\arguments{
\item{df}{a data frame or data frame extension (e.g. tibble)}

\item{...}{numeric vectors (dependent variables) for which the statistics will be calculated}

\item{IV}{grouping variable (independent variable)}

\item{spss.lab}{imports variable labels from haven labelled object (default set to TRUE)}

\item{labels.}{a character vector containing labels for DVs of user's choice}

\item{type}{type of summary statistics. Possible choices: \code{"mean_sd"}, \code{"median_iqr"}, and \code{"mrank_median"}}
}
\description{
\code{conditions_stats()} calculates a set of two basic statistics separately for each of the conditions provided by the
independent variable (IV) and for each given dependent variable (DV). The results are presented in a data.frame (tibble)
such that it is easily malleable for APA formatted publication.
}
