% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{cor_tab}
\alias{cor_tab}
\title{Compute correlations APA formatted table}
\usage{
cor_tab(df, X, Y, method = "auto", spss.lab = T, labels. = c(NULL, NULL))
}
\arguments{
\item{X, Y}{a pair of variables' sets to be correlated against one another}

\item{Which}{coefficient should be used. The default is to auto}
}
\value{
a data frame
}
\description{
cor_tab() runs multiple correlations. The coefficients can be either chosen manually
or automatically, based on the normality assumption criterion, which, in turn, is
verified in the dark with Shapiro-Wilk test.
}
\details{
The results are presented in an APA-like table with asteriks signalling statistical
significance. If an automatic choice of coefficients was selected, each estimate is
additionally accompanied with a superscript representing the chosen correlation
coefficient.
}
\examples{
cor_tab(mtcars, X = c(mpg, disp, hp), Y = c(drat, wt, qsec))
cor_tab(mtcars, X = c(mpg, disp, hp), Y = c(drat, wt, qsec), method = "pearson")
cor_tab(mtcars, X = c(mpg, disp, hp), Y = c(drat, wt, qsec), method = "spearman")

}
