% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics_apa.R
\name{demographics_apa}
\alias{demographics_apa}
\title{Compute APA-like table for demographic variables}
\usage{
demographics_apa(df, ..., spss.lab = T, labels. = NULL)
}
\arguments{
\item{df}{a data frame}

\item{...}{atomic vectors representing the demographic variables}

\item{spss.lab}{imports variables' labels from SPSS-labelled objects}

\item{labels.}{an optional character vector containing variables' names}
}
\value{
a data frame with 3 columns:
\itemize{
\item variables' names followed by their corresponding values
\item n
\item percents within the variable
}
}
\description{
Compute APA-like table for demographic variables
}
\examples{
demographics_apa(mtcars, am, vs)

labs <- c("Transmission", "Engine")
demographics_apa(mtcars, am, vs, labels. = "labs")

}
