% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots multi-way anova.R
\name{emm_mixed}
\alias{emm_mixed}
\title{Obtain emmeans for a mixed-design ANOVA}
\usage{
emm_mixed(data, bg, ..., spss.lab = T, labels. = NULL)
}
\description{
Obtain emmeans for a mixed-design ANOVA
}
