% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labellers.R
\name{extract_codebook}
\alias{extract_codebook}
\title{Extract a codebook from a data frame}
\usage{
extract_codebook(data)
}
\arguments{
\item{data}{A data frame containing variables with labels.}
}
\value{
A tibble containing two columns: variable name (Nazwa) and corresponding label (Etykieta).
}
\description{
This function generates a codebook from a data frame, where the variable names are matched with their corresponding labels.
}
\examples{
data <- data.frame(A = c(1, 2, 3),
                   B = c('A', 'B', 'C'))
attr(data$A, "label") <- "Numeric variable"
attr(data$B, "label") <- "Character variable"
extract_codebook(data)

}
