% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis_functions.R
\name{fa_loadings}
\alias{fa_loadings}
\title{Extract and sort factor loadings}
\usage{
fa_loadings(x)
}
\description{
It takes a psych "fa" object and extract factor loadings in a nicely formatted way.
Items are sorted based on their absolute value and the factor they load the most
}
