% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis_functions.R
\name{fa_reliability}
\alias{fa_reliability}
\title{Obtain Cronbach's Alpha for selected items}
\usage{
fa_reliability(df, ..., skala = "Zmienna")
}
\description{
Obtain Cronbach's Alpha for selected items
}
