% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_template_general.R
\name{git_setup}
\alias{git_setup}
\title{Set up a git repository in a specified path with specified files and folders}
\usage{
git_setup(path, files = NULL)
}
\arguments{
\item{path}{The path where the Git repository will be created and files will be added.}

\item{files}{A character vector of file names to be added to the repository.}

\item{folders}{A character vector of folder names to be added to the repository.}
}
\description{
This function initializes a new Git repository in a specified directory and adds the specified files and folders to the repository.
}
\examples{
git_setup(path = "~/myproject", files = c("script.R", "data.csv"), folders = "docs")
}
