% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_forms.R
\name{group_questions}
\alias{group_questions}
\title{Group questions in a data frame}
\usage{
group_questions(df, clean_pattern = "\\\\s\\\\*$")
}
\arguments{
\item{df}{A data frame containing a character 'question' and list 'answers' columns.}

\item{clean_pattern}{A pattern to clean the 'question' column using stringr::str_remove().}
}
\value{
A tibble with unique questions and their aggregated list of answers.
}
\description{
This function groups questions in a data frame by aggregating their corresponding answers together.
}
\examples{
df <- data.frame(question = c(rep("Q1", 3), rep("Q2", 2)), answers = c("A", "B", "C", "male", "female"))
grouped_df <- group_questions(df)
print(grouped_df)

}
