% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_modifiers.R
\name{head_more}
\alias{head_more}
\title{Return more elements of an object}
\usage{
head_more(x, n = 25)
}
\arguments{
\item{x}{an R object}

\item{n}{number of elements to be printed}
}
\value{
object of the same type as x
}
\description{
print_rows is a wrapper around head with the default number of printed elements
increased
}
