% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{imodify2}
\alias{imodify2}
\title{Modify selected columns of a date.frame using a function}
\usage{
imodify2(.x, .y, .fun, ...)
}
\arguments{
\item{.x}{A data.frame to modify}

\item{.y}{A list specifying modifications for selected elements of .x}

\item{.fun}{A function to apply to selected elements of .x}

\item{...}{Additional arguments to be passed to the function}
}
\value{
A modified data.frame
}
\description{
This function modifies selected columns of a data.frame by applying a function to them with a supplemented list of arguments.
}
\details{
This function iterates over names in .y, and for each name, it applies the function .fun to the corresponding element in .x and .y and stores the result in .x.
}
\examples{
# Define the list to modify
my_list <- list(a = 1:5, b = 6:10, c = 11:15)

# Define modifications to apply
modifications <- list(a = 10, b = 20)

# Modify the list applying the modifications using the `imodify2` function
imodify2(my_list, modifications, function(x, y) x + y)

}
