% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_analysis.R
\name{mediate_own}
\alias{mediate_own}
\title{Mediate Own}
\usage{
mediate_own(dv, treat, mediator, data, boot.ci.type = c("bca", "perc"), ...)
}
\arguments{
\item{dv}{The outcome variable.}

\item{treat}{The treatment variable.}

\item{mediator}{The mediator variable.}

\item{data}{The dataset containing the variables.}

\item{boot.ci.type}{Type of bootstrap confidence interval, should be one of 'bca' (default) or 'perc'.}

\item{...}{Additional arguments to be passed to the mediation function}
}
\value{
A tidy data frame containing mediation results.
}
\description{
This function performs mediation analysis
}
