% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiple answers question - separate.R
\name{multiple_question}
\alias{multiple_question}
\title{Separate a column with multiple answers into many binary columns}
\usage{
multiple_question(
  data,
  string,
  possible_answers,
  index_names = "",
  pattern = ","
)
}
\arguments{
\item{data}{a dataframe}

\item{string}{a character vector containing multiple answers question}

\item{possible_answers}{a character vector with the set of all possible answers}

\item{index_names}{prefix name of the newly added columns}

\item{pattern}{regex-like pattern separating the answers}
}
\value{
Returns the original dataframe expanded with the columns
}
\description{
multiple_question() creates an X number of columns in a true or false forms
in which each columns represents whether a given answer was present for a given
observation.
}
