% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{obtain_counts}
\alias{obtain_counts}
\title{Obtain counts of unique values in each column of a data frame}
\usage{
obtain_counts(df, .fun = is.atomic)
}
\arguments{
\item{df}{A data frame.}

\item{.fun}{A function to select columns. Default is is.atomic.}
}
\value{
A data frame with variables, counts, percentage, and labels.
}
\description{
This function calculates the frequency count of unique values in each column of a data frame.
}
\examples{
df <- data.frame(
  x = c('a', 'b', 'a', 'c', 'a'),
  y = c(1, 2, 1, 3, 1),
  z = c('x', 'x', 'y', 'y', 'x')
)
obtain_counts(df)

}
