% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sample.R
\name{one_sample}
\alias{one_sample}
\title{Compute a table with one sample test and descriptive statistics}
\usage{
one_sample(
  df,
  ...,
  test = c("t_test", "wilcoxon"),
  mu = 0,
  type = c("mean_sd", "median_iqr", "mrank_median"),
  spss.lab = T,
  labels. = NULL,
  digits = 2
)
}
\arguments{
\item{df}{a data frame}

\item{...}{variables on which the analysis is to be performed}

\item{test}{statistical test of choice. The default is t test.}

\item{type}{a pair of descriptive statistics. The default are mean and
standard deviation}

\item{spss.lab}{imports SPSS labels. The default is TRUE.}

\item{labels.}{a character vector containing labels for the selected variables}

\item{digits}{number of decimal rounding}
}
\description{
\code{one_sample} performs multiple one sample tests and basic descriptive
statistics of choice.
}
