% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_forms.R
\name{pdf_forms_extract_questions}
\alias{pdf_forms_extract_questions}
\title{Extract questions and answers from a pdf exported string}
\usage{
pdf_forms_extract_questions(
  string,
  split_pattern = "\\n\\n",
  question_pattern = "[*]"
)
}
\arguments{
\item{string}{A character vector containing the input string to extract questions from.}

\item{split_pattern}{A regular expression used to split the string into individual lines. Default is '\n'.}

\item{question_pattern}{A regular expression pattern to identify questions in the string. Default is '^\link{0-9}+\\.'.}
}
\value{
A data frame with columns: \code{question} (string) and \code{answers} (list of strings).
}
\description{
This function takes a string input and extracts questions and corresponding answers.
It is supposed to be used with pdf versions of a google survey imported as text.
The pattern "\link{*}" works well only, if all questions have forced response option enabled.
}
\examples{
pdf_forms_extract_questions("1. What is your name?\n2. How old are you?\nA. Answer A\nB. Answer B")
}
\references{
This function is inspired by the need to extract questions and answers from PDF forms.
}
