% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_change.R
\name{percentage_change}
\alias{percentage_change}
\title{Compute percentage change}
\usage{
percentage_change(x, reference = c("previous", "first"))
}
\arguments{
\item{x}{a numeric vector}

\item{reference}{value(s) to be treated as reference for calculating the relative percentage change}
}
\value{
a numeric vector
}
\description{
percentage_change() transforms a numeric vector into percentage values of relative change. The reference can be either
always the first value of a vector or the value that is previous to the one being calculated.
}
