% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot functions.R
\name{plot_bar}
\alias{plot_bar}
\title{Create a bar plot for many dependent variables}
\usage{
plot_bar(data, IV, ..., spss.lab = TRUE, labels. = NULL, width_wrap = 20)
}
\description{
Create a bar plot for many dependent variables
}
