% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots multi-way anova.R
\name{plot_mixed}
\alias{plot_mixed}
\title{Produce a plot for a mixed design ANOVA}
\usage{
plot_mixed(data, bg, ..., width = 20, spss.lab = T, labels. = NULL)
}
\description{
Produce a plot for a mixed design ANOVA
}
