% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca_correlations.R
\name{rbcor_tab}
\alias{rbcor_tab}
\title{Obtain a rank biserial correlation table}
\usage{
rbcor_tab(
  df,
  X,
  G,
  method = "rank_biserial",
  spss.lab = T,
  labels. = c(NULL, NULL),
  R = 1000,
  na.rm = T
)
}
\arguments{
\item{X}{ordinal variables}

\item{G}{grouping binary variable}

\item{R}{number of boostrap samples}
}
\description{
Obtain a rank biserial correlation table
}
