% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_modifiers.R
\name{recode_to_factor}
\alias{recode_to_factor}
\title{Recode a numeric vector to a factor with custom labels}
\usage{
recode_to_factor(x, labels)
}
\arguments{
\item{x}{A vector to be recoded to a factor.}

\item{labels}{A character vector of labels to assign to the levels of the factor.
The order of labels should correspond to the order of numeric levels in the vector x.}
}
\value{
A factor with the vector x recoded with the labels provided.
}
\description{
This function recodes a numeric vector to a factor with custom labels provided by the user.
Numbers are automatically derived from the length of the labels vector assuming an order as in the labels vector.
}
\examples{
recode_to_factor(c(1, 2, 3), c("Low", "Medium", "High"))
}
