% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{reg_hier}
\alias{reg_hier}
\title{compute a hierarchical linear regression APA table}
\usage{
reg_hier(models, standarize = T, digits = 2, labels = NULL)
}
\description{
compute a hierarchical linear regression APA table
}
