% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{reg_tab}
\alias{reg_tab}
\title{compute a linear regression APA table}
\usage{
reg_tab(
  mod,
  labels. = NULL,
  spss.lab = FALSE,
  digits = 2,
  coef.stats = c("b", "CI"),
  decimal = T
)
}
\description{
compute a linear regression APA table
}
