% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_modifiers.R
\name{round_df}
\alias{round_df}
\title{Round numeric vectors in a data frame}
\usage{
round_df(df, ..., digits = 3)
}
\arguments{
\item{df}{a data frame.}

\item{digits}{integer indicating the decimal place of the rounding.}
}
\value{
an object of the same type as df. The general order of rows and columns is
maintained.
}
\description{
round_df() takes a data frame and rounds of all its numeric vectors to the specified
number of decimal places (default = 0).
}
