% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labellers.R
\name{row_labs}
\alias{row_labs}
\title{Add indices' labels as separate rows}
\usage{
row_labs(df, labs, nrow)
}
\arguments{
\item{df}{a data frame}

\item{labs}{a string with labels to be added}

\item{nrow}{a numeric vector providing the row numbers for the labels}
}
\description{
Add indices' labels as separate rows
}
