% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_functions.R
\name{spss_drop}
\alias{spss_drop}
\title{Drop selected labels from a haven object}
\usage{
spss_drop(x, values, replace_na = T)
}
\arguments{
\item{x}{a labelled vector}

\item{values}{numbers with which labels to be removed are associated}

\item{replace_na}{replaces true values of the vector with NA. Default to TRUE.}
}
\value{
a vector of the same class with dropped values
}
\description{
spss_drop removes chosen value labels from a haven labelled object. If set,
it also assign NAs to values of the objects that are represented by the selected labels.
}
