% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_functions.R
\name{spss_labelled}
\alias{spss_labelled}
\title{Add spss labels to a numeric vector}
\usage{
spss_labelled(x, labels, interval = c(1, 5), var.label = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{labels}{a character vector containing user defined labels. The order will
match the numerical order}

\item{interval}{minimum and maximum value of the variable's defined scale}

\item{var.label}{optional spss label for the variable}
}
\value{
a numeric vector of class haven_labelled
}
\description{
spss_labelled assigns user defined labels to a numeric vector and changes the
class of the vector to a haven_labelled object. The object can be easily exported
to a spss database format as part of a data frame object.
}
