% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_functions.R
\name{spss_labelled2}
\alias{spss_labelled2}
\title{Turns a factor into a spss labelelled numeric vector}
\usage{
spss_labelled2(x, zero_start = FALSE, var.label = NULL)
}
\arguments{
\item{x}{a factor}

\item{zero_start}{should the numerical scale start from zero. By default set
to FALSE, meaning the scale will start from 1.}

\item{var.label}{optional spss label for the variable}
}
\value{
a numeric vector of class haven_labelled
}
\description{
spss_labelled2 turns predefined factor levels to a numeric vector with spss like labels
and changes the class of the vector to a haven_labelled object. The object can be easily exported
to a spss database format as part of a data frame object.
}
