% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_functions.R
\name{spss_write}
\alias{spss_write}
\title{Export data.frame to a spss database object with current date}
\usage{
spss_write(df, path)
}
\arguments{
\item{df}{data.frame to be exported}

\item{path}{filename / path, but without the need to add ".sav"}
}
\description{
Export data.frame to a spss database object with current date
}
