% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_modifiers.R
\name{str_stat}
\alias{str_stat}
\title{Concatenate two numeric vectors}
\usage{
str_stat(x, y, .round = c(2, 2), percent = F, comma = T)
}
\arguments{
\item{x, y}{A pair of numeric vector. y is put in parenthesis}

\item{.round}{Controls the rounding of both vectors}

\item{percent}{Adds percent character in the parenthesis}

\item{comma}{Turns the decimal mark into a comma}
}
\value{
A character vector.
}
\description{
\code{str_stat} concatenates two numeric vectors into a character vector such that the
values from the second vector are in parentheses.
}
\details{
The main purpose of the function is to provide an easy and quick way to create output that is
compatible with APA table formatting standards.
}
\examples{
x <- c(10.10, 11.55, 12.80)
y <- c(85.160, 19.840, 45.401)
str_stat(x, y)
str_stat(x, y, percent = T)
str_stat(x, y, comma = F)

}
