% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{xlsx_export}
\alias{xlsx_export}
\title{Export data frames to an Excel file}
\usage{
xlsx_export(datas, sheets, path)
}
\arguments{
\item{datas}{A list of data frames to be exported to Excel sheets.}

\item{sheets}{A character vector specifying the names of the sheets.}

\item{path}{The file path where the Excel file will be saved.}
}
\value{
The function does not return a value, it saves the Excel file to the specified location.
}
\description{
This function exports multiple data frames to separate sheets in an Excel file.
}
\examples{
xlsx_export(datas = list(df1, df2), sheets = c("Sheet1", "Sheet2"), path = "data_output.xlsx")

}
