% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_comparisons.R
\name{comparison_bg1}
\alias{comparison_bg1}
\title{Compare groups with common statistical tests}
\usage{
comparison_bg1(
  .data,
  ...,
  IV,
  test = c("t_test", "u_mann", "anova", "kruskal"),
  spss.lab = TRUE,
  labels. = NULL,
  val.labs = NULL,
  iv.lab = NULL,
  type = c("mean_sd", "median_iqr", "mrank_median")
)
}
\arguments{
\item{.data}{a data.frame or tibble vector}

\item{...}{A set of dependent variables in numeric form}

\item{IV}{The independent variable that serves as the grouping variable}

\item{test}{Statistical test used for the comparisons: \code{"t_test"} (the default), \code{"u_mann"}, \code{"anova"}, or \code{"kruskal"}}

\item{spss.lab}{Imports SPSS labels (the default). Works only, if the data were imported from a SPSS sav file with
defined variable labels.}

\item{labels.}{A character vector with labels to supplant the basic object names of the vectors. The default is set to NULL.}

\item{comma}{Chooses comma (the default) as the decimal mark}
}
\value{
A data.frame in tibble format.
The columns represent means and standard deviations for each group, test's statistic, and the corresponding p-value with
effect size.
The rows represent analyzed dependent variables.
}
\description{
\code{comparison_bg1} carries out multiple between groups comparisons with one independent variable. There are four statistical
tests to choose from:
\itemize{
\item Student's t test for independent samples
\item U Mann-Whitney test
\item One-way between groups ANOVA
\item Kruskal-Wallis test
}

The function is a wrapper around rstatix statistical test functions, but with the option for multiple testing and with the
addition of means and standard deviations. Parametric tests, i.e. t test for independent samples and One-Way ANOVA, are
preceded with Levene's test. If the homogeneity of variance assumption is not met, the function automatically applies
Welch's correction.
}
\examples{
comparison_bg1(mtcars, mpg, cyl,  hp, IV = am)
comparison_bg1(mtcars, mpg, cyl,  hp, IV = am, labels. = c("Miles per gallon", "Number of cylinders", "Horsepower"))
comparison_bg1(mtcars, mpg, cyl,  hp, IV = am, test = "u_mann")
comparison_bg1(iris, Sepal.Length, Sepal.Width, Petal.Length, Petal.Width, IV = Species, test = "kruskal")
comparison_bg1(iris, Sepal.Length, Sepal.Width, Petal.Length, Petal.Width, IV = Species, test = "anova")

}
\seealso{
\code{mean_sd()} function for solely calculating means and standard deviations of two or more groups
}
