% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Compute common statistics and Shapiro-Wilk results}
\usage{
descriptives(df, ..., IV1, IV2, digits = 2)
}
\arguments{
\item{df}{a data frame object}

\item{...}{numeric vectors used for computing descriptive statistics}

\item{IV1, IV2}{vectors used for grouping of the statistics}

\item{digits}{integer indicating the number of decimal places to be used in rounding. Default set to 2.
Doesn't include the p-value.}
}
\value{
a data frame containing descriptive statistics and Shapiro-Wilk results
}
\description{
\code{descriptives()} calculates common descriptive statistics with Shapiro-Wilk test results in an APA-like
table format.
}
