% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_vectors.R
\name{extract_hypotheses}
\alias{extract_hypotheses}
\title{Extracts hypotheses from data based on original and reformulated variables}
\usage{
extract_hypotheses(data, original, reformulated, condition = "b.z.")
}
\arguments{
\item{data}{The dataset containing the hypotheses variables}

\item{original}{The name of the original hypothesis}

\item{reformulated}{The name of the reformulated hypothesis}

\item{condition}{The condition to check for in the reformulated variable (default: 'b.z.')}
}
\value{
A named vector of hypotheses extracted from the data
}
\description{
This function extracts hypotheses from a dataset based on original and reformulated versions of the former.
The function checks for a condition in the reformulated variable and if it is met,
it replaces the original version of a hypothesis with the reformulated one.
}
\examples{
data <- tibble::tibble(
  original = c('H1', 'H2', 'H3'),
  reformulated = c('H1', 'b.z.', 'H3')
)
extract_hypotheses(data, original, reformulated, condition = 'b.z.')

}
