% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_forms.R
\name{grouped_to_list}
\alias{grouped_to_list}
\title{Convert Dataframe with aggregated answers to a ;ist}
\usage{
grouped_to_list(df, codebook, by = c(question = "Etykieta"))
}
\arguments{
\item{df}{The input dataframe containing the data to be summarized.}

\item{codebook}{The reference dataframe containing the codebook information.}

\item{by}{variables by which the join will be performed.}
}
\value{
A named list with question as names and aggregated answers as values.
}
\description{
This function takes a dataframe, codebook, and group variable, and summarizes the data by grouping based on the specified variable.
}
\examples{
df <- data.frame(Nazwa = c('A', 'A', 'B', 'B'),
                 answers = list('yes', 'no', 'good', 'bad'))
codebook <- data.frame(Etykieta = c('A', 'B'),
                       Nazwa = c('Question A', 'Question B'))
grouped_to_list(df, codebook)
}
