% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_template_spss.R
\name{make_project_spss}
\alias{make_project_spss}
\title{Make SPSS-data based analysis Project Template}
\usage{
make_project_spss(path)
}
\arguments{
\item{path}{The path to the new project.}
}
\description{
Create a new project template in RStudio which can be used for SPSS-data based analyses and simple empirical research.
}
\examples{
try(make_project_spss("path/to/new/project"))
}
