% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_analysis.R
\name{mediate_own}
\alias{mediate_own}
\title{Perform simple mediation analysis}
\usage{
mediate_own(dv, treat, mediator, data)
}
\arguments{
\item{dv}{Dependent variable.}

\item{treat}{Treatment variable.}

\item{mediator}{Mediator variable.}

\item{data}{Data frame containing the variables.}
}
\value{
A data frame with the results of total, mediation, and outcome models, along with mediation model results.
}
\description{
mediate_own calculates mediation analysis with bootstrapped confidence intervals and total, mediation, and outcome models.
}
