% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_processing.R
\name{question_labels}
\alias{question_labels}
\title{Extract question labels from a data frame with questionnaire data}
\usage{
question_labels(df, suffix = "\\\\d+")
}
\arguments{
\item{df}{A data frame}

\item{suffix}{A regular expression indicating the suffix that needs to be removed from the question labels.
The default value is '\\d+' which matches any digit.}
}
\value{
A data frame with unique values for each column, grouped by column name
}
\description{
This function takes a data frame as input and extracts unique values for each column, grouping them by the suffix column name.
It assumes a structure in which there are consecutive columns with the same prefix and a number as a suffix, e.g., "Q1", "Q2", "Q3".
There can be more than one prefix pattern.
If the data frame contains more than one type of data, all values are coerced to character type.
}
\examples{
df <- data.frame(A = c(1, 2, 3), B = c('a', 'b', 'c'))
question_labels(df)
}
