% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_summaries.R
\name{summarise_freqs}
\alias{summarise_freqs}
\title{Summarise frequency counts for multiple variables}
\usage{
summarise_freqs(data, ...)
}
\arguments{
\item{data}{A data frame containing the variables of interest.}

\item{...}{Select columns to calculate frequency counts. Possible to apply dplyr logic like 'starts_with("var")'.}
}
\value{
A nested tibble with frequency counts, percentages, and levels for each selected variable.
}
\description{
This function calculates frequency counts and percentages for multiple variables in a dataset.
}
\examples{
data(mtcars)
summarise_freqs(mtcars, cyl, gear)

}
